revno: 648 message:   Changed add/edit orgunit group from a tree view to available/selected lists view.
committer: Lars Helge Oeverland <larshelge@gmail.com>
branch nick: trunk
timestamp: Wed 2009-09-09 08:36:54 +0200

+++ dhis-2/dhis-web/dhis-web-maintenance/dhis-web-maintenance-datadictionary/src/main/java/org/hisp/dhis/dd/action/dataelementgroup/GetDataElementGroupMembersAction.java
@@ -120,7 +120,7 @@ public class GetDataElementGroupMembersAction
 
         if ( id != null )
         {
-            DataElementGroup group = dataElementService.getDataElementGroup( id.intValue() );
+            DataElementGroup group = dataElementService.getDataElementGroup( id );
 
             groupMembers = new ArrayList<DataElement>( group.getMembers() );
